Allow debugging over uart channel 0
The DEBUG_ENABLED macro is used as both an enable and the uart channel to use
for debug messages. DEBUG_ENABLED == 0 was used to denote that debugging was
disabled, but this doesn't allow for using channel 0 for debug traffic.

Instead change DEBUG_ENABLED == -1 to disable debug traffic, and a value of
0 denotes that debugging is enabled on uart channel 0.