Update the copy of examples in libraries/LINX/examples.
Update the copy of examples in libraries/LINX/examples to match
core/examples. (It was out of sync with commit e370d98b673).

Perhaps we should just remove these redundant files from git (needed only because
Arduino tools require a flat library hierarchy), add a makefile
rule to copy them here when building Arduino targets, and add the files
in this location to a .gitignore.