Fix UART on RPi3 and RPi4
The LINX UART implementation uses /dev/ttyAMA0.  This works on RPi2 but
for RPi3 and 4 they repurposed that port and the UART pins map to
/dev/ttyS0.

The RPi folks already saw this as an issue so now there is a symlink
that always points to the correct serial device (/dev/serial0).  Change
the LINX implementation to use the symlink.