Fix Device Type parsing
Previously we had been gathering the CPU type from /proc/cpuinfo, but
the format changed and we're no longer able to get CPU type from there.

Frankly, the CPU type is not a very useful piece of information,
so instead let's use that field to store the device tree model name.
The device tree model is a much more useful human-readable string (for
instance "Raspberry Pi 5 Model B").

Also modify the target config dialog to display this field as
"Hardware" instead of "CPU".