Replace Target Config .NET SSH backend with OpenSSH
When the Target Config dialog was created there was no built-in SSH
support in LabVIEW, so a .NET SSH DLL was used [1].  This .NET solution
in LV never worked very well and had intermittent issues.

Now we have an SSH (and SFTP) client built in to LabVIEW, so port the
dialog over to use this client to hopefully improve stability and
functionality.

[1] https://github.com/sshnet/SSH.NET