Make Target Config dialog handle multiple packages
When there are multiple packages of the same major version in the feed
the Target Config dialog is unable to handle it and usually shows an
empty string in the Available Version field.

This scenario is more likely now that we will have 32-bit and 64-bit ARM
packages for each newly released version.

Modify Check Target LV Package Version.vi to handle more than one
package.